LayoutTests/http/tests/inspector-enabled/console-clear-arguments-on-frame-remove-expected.txt

CONSOLE MESSAGE: line 2: A message with first argument string
CONSOLE MESSAGE: line 3: 2011
CONSOLE MESSAGE: line 4: [object Window]
Tests that console message arguments will be cleared when iframe where the messages were created is removed.

A message with first argument string resources/console-clear-arguments-iframe.html:2
2011 resources/console-clear-arguments-iframe.html:3
[object Window] resources/console-clear-arguments-iframe.html:4
